# Improve performance by reducing array slices and RegExp recreation