--- a/test/runtime/utils.test.js
+++ b/test/runtime/utils.test.js
@@ -31,4 +31,30 @@ describe('utils', () => {
       expect(res).toEqual({ some: 'thing' });
     });
   });
+
+  describe('#deepFind', () => {
+    it('finds value for a basic path', () => {
+      const obj = { a: { b: { c: 1 } } };
+      const value = utils.deepFind(obj, 'a.b.c');
+      expect(value).toEqual(1);
+    });
+
+    it('finds no value for a non-existent path', () => {
+      const obj = { a: { b: { c: 1 } } };
+      const value = utils.deepFind(obj, 'a.b.d');
+      expect(value).toEqual(undefined);
+    });
+
+    it('finds value for a key that has a dot', () => {
+      const obj = { a: { 'b.b': { c: 1 } } };
+      const value = utils.deepFind(obj, 'a.b.b.c');
+      expect(value).toEqual(1);
+    });
+
+    it('finds value for an array index', () => {
+      const obj = { a: [{ c: 1 }] };
+      const value = utils.deepFind(obj, 'a.0.c');
+      expect(value).toEqual(1);
+    });
+  });
 });