@hsource 2024-01-28
I just pushed a fix to the linter error for using `for..of`: I went back to using `.forEach` which does the same thing. Take a look when you're free!